requirements.txt
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-huggingface>=0.1.0
llama-index-readers-web>=0.1.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0